Move all tests to test directory
