+++ tests/context.py
@@ -8,3 +8,5 @@ sys.path.insert(0, os.path.abspath('.'))
 from aiutils.tftools import layers
 from aiutils.tftools import images
 from aiutils.tftools import placeholder_management
+from aiutils.tftools import batch_normalizer
+from aiutils.tftools import var_collect
