+++ tests/test_tftools.py
@@ -6,6 +6,8 @@ import tensorflow as tf
 from context import layers
 from context import images
 from context import placeholder_management
+from context import batch_normalizer
+from context import var_collect
 
 
 def test_full():
@@ -25,6 +27,9 @@ def test_full():
     y_ = np.float32(np.zeros(output_shape))
     y_hat = sess.run(y, feed_dict={x: x_})
     assert (np.all(y_hat == y_))
+
+    sess.close()
+    tf.reset_default_graph()
 
 
 def test_conv2d():
@@ -48,6 +53,9 @@ def test_conv2d():
     y_hat = sess.run(y, feed_dict={x: x_})
     assert (np.all(y_hat == y_))
 
+    sess.close()
+    tf.reset_default_graph()
+
 
 def test_batch_norm_2d():
     batch = 1
@@ -57,7 +65,7 @@ def test_batch_norm_2d():
     input_shape = [batch, in_dim]
 
     x = tf.placeholder(tf.float32, input_shape)
-    training = tf.placeholder(tf.bool, []) 
+    training = tf.placeholder(tf.bool, [])
     y = layers.batch_norm(x, training)
     sess = tf.Session()
     sess.run(tf.initialize_all_variables())
@@ -66,6 +74,9 @@ def test_batch_norm_2d():
     y_hat = sess.run(y, feed_dict={x: x_, training: True})
 
     assert y_hat.shape == x_.shape
+
+    sess.close()
+    tf.reset_default_graph()
 
 
 def test_batch_norm_4d():
@@ -78,7 +89,7 @@ def test_batch_norm_4d():
     input_shape = [batch, width, height, in_dim]
 
     x = tf.placeholder(tf.float32, input_shape)
-    training = tf.placeholder(tf.bool, []) 
+    training = tf.placeholder(tf.bool, [])
     y = layers.batch_norm(x, training)
     sess = tf.Session()
     sess.run(tf.initialize_all_variables())
@@ -87,6 +98,8 @@ def test_batch_norm_4d():
     y_hat = sess.run(y, feed_dict={x: x_, training: True})
 
     assert y_hat.shape == x_.shape
+    sess.close()
+    tf.reset_default_graph()
 
 
 def test_batch_norm_3d():
@@ -98,9 +111,94 @@ def test_batch_norm_3d():
     input_shape = [batch, width, in_dim]
 
     x = tf.placeholder(tf.float32, input_shape)
-    training = tf.placeholder(tf.bool, []) 
+    training = tf.placeholder(tf.bool, [])
     with pytest.raises(ValueError):
         y = layers.batch_norm(x, training)
+
+    tf.reset_default_graph()
+
+
+def test_batchnorm_train_mode():
+    batch = 5
+    width = 2
+    height = 3
+    channels = 4
+
+    input_shape = [batch, width, height, channels]
+
+    g = tf.Graph()
+    with g.as_default():
+        training = tf.placeholder(tf.bool, [])
+        x = tf.placeholder(tf.float32, input_shape)
+        bn = batch_normalizer.BatchNorm(x, training, name='bn')
+        y = bn.output
+        ema_mean, ema_var = bn.get_ema_moments()
+        initializer = tf.initialize_all_variables()
+
+    x_val1 = np.ones(input_shape, dtype=np.float32)
+    x_val2 = 2.0 * x_val1
+
+    sess = tf.Session(graph=g)
+    with sess.as_default():
+        sess.run(initializer)
+        y_eval1 = y.eval(feed_dict={x: x_val1, training: True})
+        ema_mean_eval1 = ema_mean.eval()
+
+        y_eval2 = y.eval(feed_dict={x: x_val2, training: True})
+        ema_mean_eval2 = ema_mean.eval()
+
+    sess.close()
+
+    assert_str = 'batch mean and var are not used correctly' + \
+                 'during training with batch norm'
+    assert (np.all(y_eval1 == np.zeros(input_shape))), assert_str
+    assert_str = 'batch mean and var are not used correctly' + \
+                 'during training with batch norm'
+    assert (np.all(y_eval2 == np.zeros(input_shape))), assert_str
+    assert_str = 'ema mean is not updated during training with batch norm'
+    assert (not np.all(ema_mean_eval1 == ema_mean_eval2)), assert_str
+
+
+def test_batchnorm_test_mode():
+    batch = 5
+    width = 2
+    height = 3
+    channels = 4
+
+    input_shape = [batch, width, height, channels]
+
+    g = tf.Graph()
+    with g.as_default():
+        training = tf.placeholder(tf.bool, [])
+        x = tf.placeholder(tf.float32, input_shape)
+        bn = batch_normalizer.BatchNorm(x, training, name='bn')
+        y = bn.output
+        ema_mean, ema_var = bn.get_ema_moments()
+        initializer = tf.initialize_all_variables()
+
+    x_val1 = np.ones(input_shape, dtype=np.float32)
+    x_val2 = 2.0 * x_val1
+
+    sess = tf.Session(graph=g)
+    with sess.as_default():
+        sess.run(initializer)
+
+        y_eval1 = y.eval(feed_dict={x: x_val1, training: False})
+        ema_mean_eval1 = ema_mean.eval()
+
+        y_eval2 = y.eval(feed_dict={x: x_val2, training: False})
+        ema_mean_eval2 = ema_mean.eval()
+
+    sess.close()
+
+    assert_str = 'ema mean and var are not used correctly' + \
+                 'during testing with batch norm'
+    assert (not np.all(y_eval1 == np.zeros(input_shape))), assert_str
+    assert_str = 'ema mean and var are not used correctly' + \
+                 'during testing with batch norm'
+    assert (not np.all(y_eval2 == np.zeros(input_shape))), assert_str
+    assert_str = 'ema mean is updated during testing with batch norm'
+    assert (np.all(ema_mean_eval1 == ema_mean_eval2)), assert_str
 
 
 def test_resize_image_like():
@@ -123,9 +221,11 @@ def test_resize_image_like():
     sess = tf.Session()
     sess.run(tf.initialize_all_variables())
     x_ = np.float32(np.random.rand(*input_shape))
-    #batch, width_height, width_height, channel))
     x_out = sess.run(x_resize, {x: x_})
     assert x_out.shape == s.get_shape()
+
+    sess.close()
+    tf.reset_default_graph()
 
 
 def test_placeholder_management():
@@ -168,3 +268,42 @@ def test_placeholder_management():
 
     assert_str = '__getitem__ method of PlaceholderManager class failed'
     assert ('word_embed' in plh_mgr['word_embed'].name), assert_str
+
+    tf.reset_default_graph()
+
+
+def test_var_collect():
+    g = tf.Graph()
+    with g.as_default():
+        with tf.name_scope('scope1') as scope1:
+            a = tf.Variable(tf.constant(1.0, shape=[1]), name='a')
+            b = tf.Variable(tf.constant(2.0, shape=[1]), name='b')
+        with tf.name_scope('scope2') as scope2:
+            c = tf.Variable(tf.constant(3.0, shape=[1]), name='c')
+
+    vars_g = var_collect.collect_all(graph=g)
+    vars_g_scope1 = var_collect.collect_scope('scope1', graph=g)
+    var_g_scope1_a = var_collect.collect_name('scope1/a', graph=g)
+    var_g_dict = var_collect.collect_list(['scope1/a', 'scope2/c'], graph=g)
+
+    # 'in' is used instead of '==' because variables have device id as well
+    list_cmp = lambda var_name_list, var_obj_list: \
+        all([v[0] in v[1].name for v in zip(var_name_list, var_obj_list)])
+
+    all_vars = ['scope1/a', 'scope1/b', 'scope2/c']
+    assert (list_cmp(all_vars, vars_g)), 'collect_all failed'
+    assert (list_cmp(['scope1/a', 'scope1/b'], vars_g_scope1)), \
+        'collect_scope failed'
+    assert (list_cmp(['scope1/a'], [var_g_scope1_a])), 'collect_name failed'
+
+    dict_check = lambda var_dict: \
+        all([key in value.name for key, value in var_dict.items()])
+    assert (dict_check(var_g_dict)), 'collect_list failed'
+
+    # Default Graph
+    d = tf.Variable(tf.constant(4.0, shape=[1]), name='d')
+    e = tf.Variable(tf.constant(5.0, shape=[1]), name='e')
+
+    def_graph_vars = var_collect.collect_all()
+    assert (list_cmp(['d', 'e'], def_graph_vars)), \
+        'collect_all failed on default graph'
